+++ tests/status_response/test_java.py
@@ -1,8 +1,8 @@
 import pytest
 
 from mcstatus.motd import Motd
-from mcstatus.responses import JavaStatusPlayer, JavaStatusPlayers, JavaStatusResponse, JavaStatusVersion
-from tests.responses import BaseStatusResponseTest
+from mcstatus.status_response import JavaStatusPlayer, JavaStatusPlayers, JavaStatusResponse, JavaStatusVersion
+from tests.status_response import BaseStatusResponseTest
 
 
 @BaseStatusResponseTest.construct
